3681.   [port]          Update the Windows build system to support feature                         selection and WIN64 builds.  This is a work in                         progress. [RT #34160]
(cherry picked from commit c3c8823fed039b3a2b8e5ca8bc2f3301d1dd840e)

Conflicts:
	CHANGES
	bin/check/win32/checktool.dsp.in
	bin/dnssec/win32/dnssectool.dsp.in
	bin/dnssec/win32/importkey.dsp.in
	bin/dnssec/win32/importkey.mak.in
	bin/named/geoip.c
	bin/named/include/named/geoip.h
	bin/tools/win32/rrchecker.dsp.in
	bin/tools/win32/rrchecker.mak.in
	config.h.win32
	lib/dns/geoip.c
	lib/dns/master.c
	lib/dns/win32/libdns.dsp.in
	lib/dns/win32/libdns.mak.in
	lib/isc/mem.c
	lib/isc/stats.c
	lib/isc/win32/file.c
	lib/isc/win32/libisc.def.in
	lib/isc/win32/libisc.mak.in
	lib/isc/win32/stdio.c
	lib/isccc/cc.c
	win32utils/BuildAll.bat
	win32utils/BuildSetup.bat
	win32utils/legacy/BINDBuild.dsw.in
	win32utils/makeversion.pl
	win32utils/setpk11provider.pl
	win32utils/updatelibxml2.pl
	win32utils/win32-build.txt

Conflicts:
	bin/dnssec/win32/verify.dsp.in
	bin/dnssec/win32/verify.mak.in
	bin/named/server.c
	bin/python/dnssec-checkds.py.in
	bin/python/dnssec-coverage.py.in
	lib/dns/nsec.c
	lib/dns/nsec3.c
	lib/dns/win32/libdns.mak.in
	lib/dns/zone.c
	lib/isc/win32/libisc.mak.in
	win32utils/BuildAll.bat
	win32utils/BuildSetup.bat
	win32utils/legacy/BINDBuild.dsw.in
	win32utils/updatelibxml2.pl
	win32utils/win32-build.txt

+++ lib/dns/gssapi_link.c
@@ -47,7 +47,7 @@
 
 #define GBUFFER_TO_REGION(gb, r) \
 	do { \
-		(r).length = (gb).length; \
+	  (r).length = (unsigned int)(gb).length; \
 		(r).base = (gb).value; \
 	} while (0)
 
@@ -180,7 +180,7 @@ gssapi_sign(dst_context_t *dctx, isc_buffer_t *sig) {
 	 * Copy the output into our buffer space, and release the gssapi
 	 * allocated space.
 	 */
-	isc_buffer_putmem(sig, gsig.value, gsig.length);
+	isc_buffer_putmem(sig, gsig.value, (unsigned int)gsig.length);
 	if (gsig.length != 0U)
 		gss_release_buffer(&minor, &gsig);
 
@@ -286,7 +286,7 @@ gssapi_destroy(dst_key_t *key) {
 static isc_result_t
 gssapi_restore(dst_key_t *key, const char *keystr) {
 	OM_uint32 major, minor;
-	size_t len;
+	unsigned int len;
 	isc_buffer_t *b = NULL;
 	isc_region_t r;
 	gss_buffer_desc gssbuffer;
@@ -346,13 +346,13 @@ gssapi_dump(dst_key_t *key, isc_mem_t *mctx, char **buffer, int *length) {
 		gss_release_buffer(&minor, &gssbuffer);
 		return (ISC_R_NOMEMORY);
 	}
-	isc_buffer_init(&b, buf, len);
+	isc_buffer_init(&b, buf, (unsigned int)len);
 	GBUFFER_TO_REGION(gssbuffer, r);
 	result = isc_base64_totext(&r, 0, "", &b);
 	RUNTIME_CHECK(result == ISC_R_SUCCESS);
 	gss_release_buffer(&minor, &gssbuffer);
 	*buffer = buf;
-	*length = len;
+	*length = (int)len;
 	return (ISC_R_SUCCESS);
 }
