+++ lib/dns/ssu_external.c
@@ -131,7 +131,7 @@ dns_ssu_external_match(dns_name_t *identity,
 	isc_buffer_t *tkey_token = NULL;
 	int fd;
 	const char *sock_path;
-	size_t req_len;
+	unsigned int req_len;
 	isc_region_t token_region;
 	unsigned char *data;
 	isc_buffer_t buf;
