+++ lib/isc/hash.c
@@ -94,7 +94,7 @@ struct isc_hash {
 	isc_boolean_t	initialized;
 	isc_refcount_t	refcnt;
 	isc_entropy_t	*entropy; /*%< entropy source */
-	unsigned int	limit;	/*%< upper limit of key length */
+	size_t		limit;	/*%< upper limit of key length */
 	size_t		vectorlen; /*%< size of the vector below */
 	hash_random_t	*rndvector; /*%< random vector for universal hashing */
 };
@@ -140,7 +140,7 @@ static unsigned char maptolower[] = {
 
 isc_result_t
 isc_hash_ctxcreate(isc_mem_t *mctx, isc_entropy_t *entropy,
-		   unsigned int limit, isc_hash_t **hctxp)
+		   size_t limit, isc_hash_t **hctxp)
 {
 	isc_result_t result;
 	isc_hash_t *hctx;
@@ -250,7 +250,8 @@ isc_hash_ctxinit(isc_hash_t *hctx) {
 		isc_result_t result;
 
 		result = isc_entropy_getdata(hctx->entropy,
-					     hctx->rndvector, hctx->vectorlen,
+					     hctx->rndvector,
+					     (unsigned int)hctx->vectorlen,
 					     NULL, 0);
 		INSIST(result == ISC_R_SUCCESS);
 #else
@@ -258,7 +259,7 @@ isc_hash_ctxinit(isc_hash_t *hctx) {
 #endif
 	} else {
 		isc_uint32_t pr;
-		unsigned int i, copylen;
+		size_t i, copylen;
 		unsigned char *p;
 
 		p = (unsigned char *)hctx->rndvector;
